+++ packages/frontend/src/components/MkSwitch.vue
@@ -70,6 +70,10 @@ const toggle = () => {
 	&.disabled {
 		opacity: 0.6;
 		cursor: not-allowed;
+
+		.label {
+			cursor: not-allowed;
+		}
 	}
 }