+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ## 2026.5.0
 
 ### General
+- Feat: Misskey Gamesのプレイ可否をロールで設定可能に
 - Enhance: アバターデコレーションにカテゴリを設定できるように
 
 ### Client